feat: 관심주식 페이지 컴포넌트 일부 서버 컴포넌트 변경
